+++ lib/Perl/Tidy.pm
@@ -111,7 +111,7 @@ BEGIN {
     # then the Release version must be bumped, and it is probably past time for
     # a release anyway.
 
-    $VERSION = '20230701.02';
+    $VERSION = '20230701.03';
 } ## end BEGIN
 
 sub DESTROY {