+++ lib/Perl/Tidy/IOScalar.pm
@@ -10,7 +10,7 @@ package Perl::Tidy::IOScalar;
 use strict;
 use warnings;
 use Carp;
-our $VERSION = '20230701.02';
+our $VERSION = '20230701.03';
 
 use constant EMPTY_STRING => q{};