+++ lib/Perl/Tidy/IndentationItem.pm
@@ -8,7 +8,7 @@
 package Perl::Tidy::IndentationItem;
 use strict;
 use warnings;
-our $VERSION = '20230701.02';
+our $VERSION = '20230701.03';
 
 BEGIN {